+++ src/miner/utils/src/DataType.cc
@@ -14,14 +14,14 @@ void DataType::setName(const std::string &name) { _name = name; }
 
 std::string DataType::getName() const { return _name; }
 
-void DataType::setType(VarType type, uint8_t size) {
+void DataType::setType(VarType type, size_t size) {
     _type = type;
     _size = size;
 }
 
 VarType DataType::getType() const { return _type; }
 
-uint8_t DataType::getSize() const { return _size; }
+size_t DataType::getSize() const { return _size; }
 
 void DataType::setLocation(Location dir) { _location = dir; }
